Add Support of GitFlow branching Model
Add support for the GitFlow branching Model in the Travis CI/CD Script
Only Releases and Release Candidates are deployed to central Repository
by providing a Tag like 1.2.3 or 1.2.3-RC3.
Snapshots are deployed to central Repository too, but only when they
are in develop branch and triggered by a commit.
For all other jobs only the CI part is done.

Fix #1